Modified Setup.txt
Apart from a couple of bug-fixes to existing functions, added two new Setup functions for setting up tables and multiple divs within a div.
The new CreateTableSetupWithClickFn is very similar to the existing table setup function except instead of navigating to a subpage,
it will an arbitrary generic function.
The motivation for this function was to allow a form dialog to be populated and shown.
The new CreateMultiDivSetupFn is similar to the table setup function but adds new DIV child elements to parent div for each result row.

+	Modified Setup.CreateTableSetupWithDownFn - bug fix: fixed determination of 'limit'.
+	Modified Setup.CreateTableSetupWithDownFn - modified code executed if zero rows are returned to now call Setup.Clear (new function)
	                                            and append new "No entries added." table row to prevent deleting any template elements.
+	Added Setup.CreateTableSetupFn.RetrieveTemplateParameters function.
+	Added Setup.CreateTableSetupWithClickFn function.
+	Added Setup.CreateMultiDivSetupFn function.
+	Added Setup.More function.